requirements.txt
python-telegram-bot==21.0.1
google-generativeai
python-dotenv